Add 'date' and 'layout' to the front-matter of Forestry pages
Some pages rely on the defaults in _config.yml for their 'layout' value.
Some pages also don't require a 'date' in their front-matter. This works
for Jekyll.

However, if those pages were ever edited with Forestry, Forestry will
generate `date: ''` and `layout: ''` in the front-matter which will
override the default values. Having an empty string as `date` value
breaks the build because it's an invalid date. Having an empty string as
`layout` value throws a warning during build and break the visuals on
the site.

To avoid problem with Forestry edits, we backfill those pages with the
actual value for `layout` and today's date for `date`.